Remove gevent support
