+++ .travis.yml
@@ -1,6 +1,4 @@
 language: python
-python:
-  - "3.5"
 dist: trusty
 sudo: required
 env:
@@ -9,16 +7,16 @@ env:
     - GETH_BASE_INSTALL_PATH=$TRAVIS_BUILD_DIR
     - CHAIN_SYNC_BLOCK_DELTA=100
     - GOROOT=/usr/lib/go-1.7
-    - PYTEST_ADDOPTS="-n 3"
+    - PYTEST_ADDOPTS="-n 2"
   matrix:
     # Solc 0.4.8
-    - SOLC_VERSION=v0.4.8 TOX_POSARGS="-e py27-stdlib -e py27-gevent -e py34-stdlib -e py34-gevent -e py35-stdlib -e py35-gevent" GETH_VERSION=v1.6.6
+    - SOLC_VERSION=v0.4.8 TOX_POSARGS="-e py27 -e py34 -e py35 GETH_VERSION=v1.6.6
     # Solc 0.4.11
-    - SOLC_VERSION=v0.4.11 TOX_POSARGS="-e py27-stdlib -e py27-gevent -e py34-stdlib -e py34-gevent -e py35-stdlib -e py35-gevent" GETH_VERSION=v1.6.6
+    - SOLC_VERSION=v0.4.11 TOX_POSARGS="-e py27 -e py34 -e py35 GETH_VERSION=v1.6.6
     # Solc 0.4.12
-    - SOLC_VERSION=v0.4.12 TOX_POSARGS="-e py27-stdlib -e py27-gevent -e py34-stdlib -e py34-gevent -e py35-stdlib -e py35-gevent" GETH_VERSION=v1.6.6
+    - SOLC_VERSION=v0.4.12 TOX_POSARGS="-e py27 -e py34 -e py35 GETH_VERSION=v1.6.6
     # Solc 0.4.13
-    - SOLC_VERSION=v0.4.13 TOX_POSARGS="-e py27-stdlib -e py27-gevent -e py34-stdlib -e py34-gevent -e py35-stdlib -e py35-gevent" GETH_VERSION=v1.6.6
+    - SOLC_VERSION=v0.4.13 TOX_POSARGS="-e py27 -e py34 -e py35 GETH_VERSION=v1.6.6
     # Linting
     - TOX_POSARGS="-e flake8"
 cache:
