+++ populus/api/compile_contracts.py
@@ -1,12 +1,11 @@
+import threading
+
 from populus.compilation import (
     compile_project_contracts,
 )
 
 from populus.utils.cli import (
     watch_project_contracts,
-)
-from populus.utils.compat import (
-    spawn,
 )
 from populus.utils.compile import (
     write_compiled_sources,
@@ -19,8 +18,10 @@ def compile_project(project, watch):
     write_compiled_sources(project.compiled_contracts_asset_path, compiled_contracts)
 
     if watch:
-        thread = spawn(
-            watch_project_contracts,
-            project=project,
+        thread = threading.Thread(
+            target=watch_project_contracts,
+            kwargs={'project': project},
         )
+        thread.daemon = True
+        thread.start()
         thread.join()
