--- requirements-gevent.txt
@@ -1,6 +1,0 @@
--r requirements-dev.txt
-gevent>=1.1.2,<1.2.0
-web3[gevent]>=3.5.1
-eth-testrpc[gevent]>=1.1.0
-py-geth[gevent]>=1.7.0
-py-solc[gevent]>=1.0.0
