+++ tests/networking-utils/test_wait_for_http_connection.py
@@ -1,7 +1,5 @@
-import pytest
-
-import os
 import random
+import threading
 
 try:
     from http.server import (
@@ -14,31 +12,26 @@ except ImportError:
         BaseHTTPRequestHandler,
     )
 
-from populus.utils.compat import (
-    spawn,
-    Timeout,
-)
 from populus.utils.networking import (
     get_open_port,
     wait_for_connection,
 )
-
-
-def can_test_connect_with_gevent():
-    if os.environ.get('THREADING_BACKEND', 'stdlib') != 'gevent':
-        return True
-    try:
-        from gevent import monkey
-    except ImportError:
-        return False
-
-    return monkey.is_module_patched('socket')
-
-
-@pytest.mark.skipif(
-    can_test_connect_with_gevent(),
-    reason="Bad configuration for gevent based testing",
+from populus.utils.wait import (
+    Timeout,
 )
+
+
+def _spawn(target, *args, **kwargs):
+    thread = threading.Thread(
+        target=target,
+        args=args,
+        kwargs=kwargs,
+    )
+    thread.daemon = True
+    thread.start()
+    return thread
+
+
 def test_wait_for_connection_success():
     success_tracker = {}
     port = get_open_port()
@@ -74,8 +67,8 @@ def test_wait_for_connection_success():
         success_tracker.setdefault('server_success', True)
         success_tracker['server_exited'] = True
 
-    spawn(_do_client)
-    spawn(_do_server)
+    _spawn(_do_client)
+    _spawn(_do_server)
 
     try:
         with Timeout(5) as _timeout:
@@ -104,7 +97,7 @@ def test_wait_for_connection_failure():
         else:
             success_tracker['client_success'] = True
 
-    spawn(_do_client)
+    _spawn(_do_client)
 
     try:
         with Timeout(5) as _timeout:
