+++ tests/observer-utils/test_watchdog_based_observer.py
@@ -1,5 +1,3 @@
-import pytest
-
 import os
 
 from populus.utils.filesystem import (
@@ -8,14 +6,9 @@ from populus.utils.filesystem import (
 from populus.utils.six import (
     queue,
 )
-from populus.utils.observers.observers_watchdog import DirWatcher
+from populus.utils.observers import DirWatcher
 
 
-def is_gevent_enabled():
-    return os.environ.get('THREADING_BACKEND', 'stdlib') == 'gevent'
-
-
-@pytest.mark.skipif(not is_gevent_enabled(), reason="gevent has been enabled")
 def test_watchdog_based_observer(temporary_dir):
     file_path_a = os.path.join(temporary_dir, 'file-a.txt')
