+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 envlist=
-    py{27,34,35}-{stdlib,gevent}
+    py{27,34,35}
     flake8
 
 [flake8]
@@ -14,11 +14,7 @@ passenv =
     SOLC_BINARY
     GETH_BINARY
     TRAVIS_BUILD_DIR
-setenv =
-    gevent: THREADING_BACKEND=gevent
-deps =
-    stdlib: -r{toxinidir}/requirements-dev.txt
-    gevent: -r{toxinidir}/requirements-gevent.txt
+deps=-r{toxinidir}/requirements-dev.txt
 basepython =
     py27: python2.7
     py34: python3.4
@@ -26,6 +22,4 @@ basepython =
 
 [testenv:flake8]
 basepython=python
-deps =
-    -r{toxinidir}/requirements-gevent.txt
 commands=flake8 {toxinidir}/populus {toxinidir}/tests
